wraper takes on requirments
