+++ lib/include/rmb/command/SelfManagedCommand.h
@@ -9,11 +9,11 @@
 namespace rmb {
   /**
    * Scheduels a command that will delete itself after `End()` is called. This 
-   * is Achived by wraping the command in an internal `SuicidalCommand` that 
-   * takes ownership of the command. This `Sucidal command is schedueld and
+   * is Achived by wraping the command in an internal `SelfManaged` that 
+   * takes ownership of the command. This `SelfManagedCommand` is schedueld and
    * will delete itself and the inital command with it once it is ended.
    * 
    * @param command Command to be schedueled (construt in place).
    **/
-  void scheduelSuicidalCommand(std::unique_ptr<frc2::Command> command);
+  void scheduelSelfManagedCommand(std::unique_ptr<frc2::Command> command);
 } // namespace rmb
